Extract header-nav component to reduce duplication
I am less convinced that this is the right solution, but it will make moving forward simpler
