+++ app/components/top-bar.js
@@ -1,6 +1,3 @@
-/* global HS */
-
-import { htmlSafe } from '@ember/string';
 import Component from '@ember/component';
 import Ember from 'ember';
 import { computed } from 'ember-decorators/object';
@@ -30,36 +27,7 @@ export default Component.extend(InViewportMixin, {
 
   @alias('broadcastsService.broadcasts') broadcasts: null,
 
-  @computed()
-  deploymentVersion() {
-    if (window && window.location) {
-      const hostname = window.location.hostname;
-
-      if (hostname.indexOf('ember-beta') === 0 || hostname.indexOf('ember-canary') === 0) {
-        return `Ember ${Ember.VERSION}`;
-      } else if (hostname.indexOf('test-deployments') > 0) {
-        const branchName = hostname.split('.')[0];
-        const branchURL = this.get('externalLinks').travisWebBranch(branchName);
-        const branchLink = `<a href='${branchURL}'><code>${branchName}</code></a>`;
-
-        return htmlSafe(`Test deployment ${branchLink}`);
-      } else {
-        return false;
-      }
-    } else {
-      return false;
-    }
-  },
-
   actions: {
-    signIn() {
-      return this.get('signIn')();
-    },
-
-    signOut() {
-      return this.get('signOut')();
-    },
-
     toggleBurgerMenu() {
       this.toggleProperty('is-open');
       return false;
@@ -74,29 +42,11 @@ export default Component.extend(InViewportMixin, {
       this.get('broadcastsService').markAsSeen(broadcast);
       return false;
     },
-
-    helpscoutTrigger() {
-      HS.beacon.open();
-      return false;
-    }
   },
 
   @computed('auth.signedIn', 'landingPage', 'features.proVersion')
   showCta(signedIn, landingPage, pro) {
     return !signedIn && !landingPage && !pro;
-  },
-
-  @computed('tab', 'auth.state')
-  classProfile(tab, authState) {
-    let classes = ['profile menu'];
-
-    if (this.get('tab') === 'profile') {
-      classes.push('active');
-    }
-
-    classes.push(authState || 'signed-out');
-
-    return classes.join(' ');
   },
 
   didInsertElement() {
