+++ app/templates/owner-loading.hbs
@@ -31,55 +31,7 @@
         {{/if}}
       {{/unless}}
 
-      <nav id="navigation" class="navigation {{if is-open 'is-open'}}">
-        <ul>
-          {{#unless features.enterpriseVersion}}
-            {{#unless features.proVersion}}
-              <li><a href="https://blog.travis-ci.com" title="Travis CI Blog" class="navigation-anchor">Blog</a></li>
-              <li><a href="https://www.traviscistatus.com/" title="Travis CI Status" class="navigation-anchor">Status</a></li>
-              <li>
-                <span class="navigation-anchor">Help</span>
-                <ul class="navigation-nested">
-                  <li><a href="https://docs.travis-ci.com" title="Browse our documentation">Read Our Docs</a></li>
-                  <li><a href="https://docs.travis-ci.com/imprint.html" title="Go to the imprint">Imprint</a></li>
-                </ul>
-              </li>
-            {{/unless}}
-
-            {{#if features.proVersion}}
-              {{#unless auth.signedIn}}
-                <li><a href="/about" title="Travis CI team" class="navigation-anchor">About Us</a></li>
-                <li><a href="/plans" title="Plans and Prices for Travis CI" class="navigation-anchor">Plans & Pricing</a></li>
-                <li><a href="https://enterprise.travis-ci.com" title="Travis CI for Enterprise" class="navigation-anchor">Enterprise</a></li>
-            {{else}}
-                <li><a href="https://www.traviscistatus.com/" title="Travis CI Status" class="navigation-anchor">Status</a></li>
-                <li>
-                  <span class="navigation-anchor">Help</span>
-                  <ul class="navigation-nested">
-                    <li><a href="mailto:support@travis-ci.com" title="Email the Travis support team" {{action 'helpscoutTrigger'}}>Email Support</a></li>
-                    <li><a href="https://docs.travis-ci.com" title="Browse our documentation">Read Our Docs</a></li>
-                    <li><a href="https://twitter.com/travisci" title="Travis in the Twitter-sphere">Twitter</a></li>
-                  </ul>
-                </li>
-                <li>
-                  <span class="navigation-anchor">Legal</span>
-                  <ul class="navigation-nested">
-                    <li><a href={{config.urls.imprint}} title="Go to the imprint">Imprint</a></li>
-                    <li><a href={{config.urls.security}} title="Go to our security statement">Security</a></li>
-                    <li><a href={{config.urls.terms}} title="Go to our terms of service">Terms</a></li>
-                  </ul>
-                </li>
-              {{/unless}}
-            {{/if}}
-
-            {{#if deploymentVersion}}<li><span class='deployment-version'>{{deploymentVersion}}</span></li>{{/if}}
-    {{else}}
-            {{#if auth.signedIn}}
-              <li><a class="navigation-anchor" title="Documentation" href="https://docs.travis-ci.com">Docs</a></li>
-            {{/if}}
-          {{/unless}}
-        </ul>
-      </nav>
+      {{header-nav}}
     </div>
   </header>
